+++ app/src/main/java/com/cornellappdev/android/volume/TabbedActivity.kt
@@ -1,0 +1,65 @@
+package com.cornellappdev.android.volume
+
+import android.os.Bundle
+import android.util.Log
+import androidx.appcompat.app.AppCompatActivity
+import androidx.viewpager.widget.ViewPager
+import com.cornellappdev.android.volume.adapters.CustomPagerAdapter
+import com.cornellappdev.android.volume.models.Article
+import com.cornellappdev.android.volume.models.Publication
+import com.google.android.material.tabs.TabLayout
+
+class TabbedActivity : AppCompatActivity() {
+
+    private lateinit var viewPager: ViewPager
+    private lateinit var tabbedLayout: TabLayout
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.tabbed_activity)
+        viewPager = findViewById(R.id.view_pager_fragments)
+        tabbedLayout = findViewById(R.id.tab_layout)
+        viewPager.adapter = CustomPagerAdapter(supportFragmentManager, tabbedLayout.tabCount)
+        tabbedLayout.setupWithViewPager(viewPager)
+
+        for (i in 0 until tabbedLayout.tabCount) {
+            if (i == 0){
+                tabbedLayout.getTabAt(i)?.setIcon(R.drawable.ic_volumesvg_orange)
+            }
+            if (i == 1){
+                tabbedLayout.getTabAt(i)?.setIcon(R.drawable.ic_book_gray)
+            }
+            if(i==2) {
+                tabbedLayout.getTabAt(i)?.setIcon(R.drawable.ic_bookmark_gray)
+            }
+        }
+
+        viewPager.addOnPageChangeListener(object : ViewPager.OnPageChangeListener {
+
+            override fun onPageScrollStateChanged(state: Int) {
+            }
+
+            override fun onPageScrolled(position: Int, positionOffset: Float, positionOffsetPixels: Int) {
+            }
+            override fun onPageSelected(position: Int) {
+                when (position) {
+                    0 -> {
+                        tabbedLayout.getTabAt(0)?.setIcon(R.drawable.ic_volumesvg_orange)
+                        tabbedLayout.getTabAt(1)?.setIcon(R.drawable.ic_book_gray)
+                        tabbedLayout.getTabAt(2)?.setIcon(R.drawable.ic_bookmark_gray)
+                    }
+                    1 -> {
+                        tabbedLayout.getTabAt(0)?.setIcon(R.drawable.ic_volumesvg_gray)
+                        tabbedLayout.getTabAt(1)?.setIcon(R.drawable.ic_book_orange)
+                        tabbedLayout.getTabAt(2)?.setIcon(R.drawable.ic_bookmark_gray)
+                    }
+                    2 ->{
+                        tabbedLayout.getTabAt(0)?.setIcon(R.drawable.ic_volumesvg_gray)
+                        tabbedLayout.getTabAt(1)?.setIcon(R.drawable.ic_book_gray)
+                        tabbedLayout.getTabAt(2)?.setIcon(R.drawable.ic_bookmark_orange)}
+                }
+            }
+
+        })
+    }
+}
