+++ app/src/main/java/com/cornellappdev/android/volume/adapters/CustomPagerAdapter.kt
@@ -1,0 +1,24 @@
+package com.cornellappdev.android.volume.adapters
+
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentPagerAdapter
+import com.cornellappdev.android.volume.fragments.HomeFragment
+import com.cornellappdev.android.volume.fragments.PublicationsFragment
+import com.cornellappdev.android.volume.fragments.SavedPublicationsFragment
+
+class CustomPagerAdapter(fragmentManager: FragmentManager, private val numOfTabs: Int)
+    : FragmentPagerAdapter(fragmentManager, BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT) {
+
+    override fun getCount(): Int {
+        return numOfTabs
+    }
+
+    override fun getItem(position: Int): Fragment {
+        return when (position) {
+            0 -> HomeFragment()
+            1 -> PublicationsFragment()
+            else -> SavedPublicationsFragment()
+        }
+    }
+}
