+++ app/src/main/java/com/cornellappdev/android/volume/adapters/OnboardingPageAdapter.kt
@@ -1,0 +1,26 @@
+package com.cornellappdev.android.volume.adapters
+
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentPagerAdapter
+import com.cornellappdev.android.volume.fragments.OnboardingFrag1
+import com.cornellappdev.android.volume.fragments.OnboardingFrag2
+import com.cornellappdev.android.volume.models.Publication
+
+private const val COUNT = 3
+
+class OnboardingPageAdapter(fragmentManager: FragmentManager, val publications : List<Publication>) : FragmentPagerAdapter(fragmentManager) {
+
+    override fun getCount(): Int {
+        return COUNT
+    }
+
+    override fun getItem(position: Int): Fragment {
+        return when (position) {
+            0 -> OnboardingFrag1()
+            else -> OnboardingFrag2(publications = publications)
+        }
+    }
+
+
+}
