--- app/src/main/java/com/example/volume_android/BookmarkActivity.kt
@@ -1,13 +1,0 @@
-package com.example.volume_android
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-
-class BookmarkActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.bookmark_activity)
-
-    }
-}
