--- app/src/main/res/layout/bookmark_activity.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <ImageView
-        android:id="@+id/bookmark_title"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        android:src="@drawable/ic_bookmarktitlesvg"
-        android:layout_marginTop="35dp"
-        android:layout_marginLeft="15dp"/>
-
-    <ImageView
-        android:id="@+id/bookmark_savedarticles"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintTop_toBottomOf="@id/bookmark_title"
-        app:layout_constraintLeft_toLeftOf="@id/bookmark_title"
-        android:src="@drawable/ic_savedarticlestitle"
-        android:layout_marginTop="40dp"/>
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        tools:listitem="@layout/home_bookmarks_article_card"
-        android:layout_marginLeft="5dp"
-        app:layout_constraintTop_toBottomOf="@id/bookmark_savedarticles"
-        app:layout_constraintLeft_toLeftOf="@id/bookmark_savedarticles"
-        android:layout_marginTop="5dp"/>
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
